RealityWeb/requirements.txt
beautifulsoup4==4.9.1
Django==3.1.1
django-crispy-forms==1.9.2
django-filter==2.3.0
django-map-widgets==0.3.0
djangorestframework==3.11.1
django-multiselectfield==0.1.12
Fiona==1.8.13
folium==0.11.0
geojson==2.5.0
geopandas==0.8.1
Jinja2==2.11.2
lxml==4.5.2
# mod-wsgi==4.7.1
numpy==1.19.1
overpass==0.7
overpy==0.4
pandas==1.1.1
# psycopg2==2.8.5
psycopg2-binary==2.8.5
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
Shapely==1.7.0
SQLAlchemy==1.3.19
urllib3==1.25.10
django-plotly-dash==1.4.2
channels==2.4.0
daphne==2.5.0
redis==3.5.3
django-redis==4.12.1
channels-redis==3.1.0